requirements.txt
torch
torchvision
opencv-python
mediapipe
rembg
torchgeometry
numpy
Pillow
gradio
